separate class and add error.csv
