+++ data_test.csv
@@ -1,0 +1,45 @@
+"2019-11-29T21:30:00+09:00","ジュマンジ","シネマシティズン（60才以上）",1000
+"2019-11-29T10:20:00+09:00","ルパン三世","シネマシティズン（60才以上）",1000
+"2019-11-29T10:00:00+09:00","スター・ウォーズ","シネマシティズン（60才以上）",1000
+"2019-11-29T22:00:00+09:00","ジョーカー","中・高校生",1000
+"2019-11-29T19:30:00+09:00","アナと雪の女王","シニア（70才以上）",1100
+"2019-11-29T21:30:00+09:00","ジュマンジ","幼児（3才以上）・小学生",1000
+"2019-11-29T22:00:00+09:00","ジョーカー","障がい者（高校以下）",900
+"2019-11-29T13:00:00+09:00","アナと雪の女王","幼児（3才以上）・小学生",1000
+"2019-11-29T21:30:00+09:00","ジュマンジ","シニア（70才以上）",1100
+"2019-11-29T22:00:00+09:00","ジョーカー","一般",1300
+"2019-11-29T13:00:00+09:00","アナと雪の女王","中・高校生",1000
+"2019-11-29T13:00:00+09:00","アナと雪の女王","シネマシティズン",1000
+"2019-11-29T19:00:00+09:00","ルパン三世","障がい者（学生以上）",1000
+"2019-11-29T21:30:00+09:00","ジュマンジ","シネマシティズン",1000
+"2019-11-29T22:00:00+09:00","ジョーカー","学生（大・専）",1300
+"2019-11-30T17:00:00+09:00","ジョーカー","シネマシティズン（60才以上）",1000
+"2019-11-30T21:30:00+09:00","ルパン三世","シネマシティズン（60才以上）",1000
+"2019-11-30T19:30:00+09:00","ルパン三世","シニア（70才以上）",1100
+"2019-11-30T13:10:00+09:00","ジョーカー","シニア（70才以上）",1100
+"2019-11-30T16:30:00+09:00","スター・ウォーズ","シネマシティズン",1300
+"2019-11-30T13:00:00+09:00","スター・ウォーズ","シニア（70才以上）",1100
+"2019-11-30T13:00:00+09:00","スター・ウォーズ","シネマシティズン",1300
+"2019-11-30T10:20:00+09:00","ルパン三世","障がい者（学生以上）",1000
+"2019-11-30T22:00:00+09:00","スター・ウォーズ","中・高校生",1000
+"2019-11-30T10:00:00+09:00","ジュマンジ","シネマシティズン（60才以上）",1000
+"2019-11-30T10:00:00+09:00","ジュマンジ","幼児（3才以上）・小学生",1000
+"2019-11-30T16:30:00+09:00","スター・ウォーズ","シネマシティズン",1300
+"2019-11-30T17:00:00+09:00","ジョーカー","学生（大・専）",1500
+"2019-11-30T16:30:00+09:00","スター・ウォーズ","中・高校生",1000
+"2019-11-30T16:30:00+09:00","スター・ウォーズ","幼児（3才以上）・小学生",1000
+"2019-12-01T17:00:00+09:00","スター・ウォーズ","学生（大・専）",1100
+"2019-12-01T22:00:00+09:00","ルパン三世","シネマシティズン",1100
+"2019-12-01T13:00:00+09:00","スター・ウォーズ","シニア（70才以上）",1100
+"2019-12-01T17:00:00+09:00","スター・ウォーズ","幼児（3才以上）・小学生",1000
+"2019-12-01T17:00:00+09:00","スター・ウォーズ","幼児（3才以上）・小学生",1000
+"2019-12-01T19:30:00+09:00","アナと雪の女王","シネマシティズン（60才以上）",1000
+"2019-12-01T13:00:00+09:00","スター・ウォーズ","シネマシティズン（60才以上）",1000
+"2019-12-01T21:30:00+09:00","ルパン三世","シネマシティズン（60才以上）",1000
+"2019-12-01T13:00:00+09:00","スター・ウォーズ","幼児（3才以上）・小学生",1000
+"2019-12-01T16:30:00+09:00","ジュマンジ","学生（大・専）",1100
+"2019-12-01T19:00:00+09:00","スター・ウォーズ","学生（大・専）",1100
+"2019-12-01T10:20:00+09:00","ジョーカー","中・高校生",1000
+"2019-12-01T16:30:00+09:00","ジュマンジ","シニア（70才以上）",1100
+"2019-12-01T16:30:00+09:00","ジュマンジ","幼児（3才以上）・小学生",1000
+"2019-12-01T10:00:00+09:00","アナと雪の女王","一般",1100
