+++ error.csv
@@ -1,0 +1,4 @@
+(line:1)「テスト料金タイプ」は存在しない料金タイプです)
+(line:1)「テストタイトル」は上映されていない作品です
+(line:2)「未公開タイトル」は上映されていない作品です
+(line:3)「お客さん」は存在しない料金タイプです)
