+++ result.csv
@@ -1,0 +1,20 @@
+▼サマリー
+売上： 48,100円
+
+▼作品別売上
+・ジュマンジ: 9,300円
+・ルパン三世: 7,200円
+・スター・ウォーズ: 16,300円
+・ジョーカー: 9,100円
+・アナと雪の女王: 6,200円
+
+▼料金タイプ別売上
+・シネマシティズン（60才以上）: 9,000円
+・中・高校生: 5,000円
+・シニア（70才以上）: 7,700円
+・幼児（3才以上）・小学生: 8,000円
+・障がい者（高校以下）: 900円
+・一般: 2,400円
+・シネマシティズン: 7,000円
+・障がい者（学生以上）: 2,000円
+・学生（大・専）: 6,100円
